Add callout_notify function
A notifier is introduced to invoke notification scripts that can signal
other programs of mdevctl's retults.

The notify callout is invoked once at the tail-end of mdevctl to signal
other programs of the mdevctl command's results. Errors reported during
the notify callout are non-disruptive.

The notification scripts are stored in /etc/mdevctl.d/scripts.d/notifiers diretorry.
All scripts in that directory will be executed.

Signed-off-by: Zhuoying Cai <zycai@linux.vnet.ibm.com>
Reviewed-by: Collin Walling <walling@linux.ibm.com>