Add callout get function
Add "get" callout for mdevctl to request extraneous data to be retrieved
by the script.

Unlike the curently existing callout function, which will only validate
the device setups. This form of callout can be used
1) during mdevctl list --dumpjson and
2) during mdevctl define -u <uuid>
to gather the attributes of an active device from a callout script.

Unlike the other callouts, this one expects a JSON-formatted array of
attributes to be returned on stdout.

Example of expected output from a script:
[
    {
        "assign_adapter": "2"
    },
    {
        "assign_domain": "1"
    }
]

Signed-off-by: Zhuoying Cai <zycai@linux.vnet.ibm.com>
Reviewed-by: Collin Walling <walling@linux.ibm.com>