Add callout function to mdevctl
Add callout functionality to invoke external scripts that are designed to extend
the mdevctl command operations for specific device types. There are two types
of callouts: pre-command callout, which occurs prior to "primary command
execution" (such as prior to writing the device config during a define command),
and post-command callout, which occurs after the "primary command execution".

The pre-command callout is invoked before any changes are made to the device
or device configuration. A script may check if a device's configuration is sane
before mdevctl commits any changes. Errors reported during the pre-command
callout are disruptive -- mdevctl will exit early if an error is detected, and no
change to the device configuration will be committed.

The post-command callout is invoked after changes have been made to the
device or device configuration. The script may check if the device / system
environment is sane after a device is started/configured, or perform cleanups
if the primary command execution fails. Errors reported during the post-command
callout are non-disruptive.

Both callouts will be given an event type (to denote the callout type), an action
(corresponding to an mdevctl command), a command status (none if mdevctl
has not yet performed the "primary command execution", or success/failure
depending on result of the command), along with the device's type, uuid, and
parent. The device's JSON-formatted configuration file will also be provided
via standard input.

The callout scripts are stored in /etc/mdevctl.d/scripts.d/callouts directory.
Each script in the directory will be attempted until one matches the device-type.
If no scripts are found, the call-out is skipped and mdevctl will carry on as normal.

Signed-off-by: Zhuoying Cai <zycai@linux.vnet.ibm.com>
Reviewed-by: Collin Walling <walling@linux.ibm.com>